(int)
git-svn-id: svn://136.177.114.72/svn_GW/phreeqc3/trunk@6909 1feff8c3-07ed-0310-ac33-dd36852eb9cd

+++ src/phreeqc/inverse.cpp
@@ -3964,7 +3964,7 @@ dump_netpath(struct inverse *inverse_ptr)
 		/* total number of wells */
 		fprintf(netpath_file,
 				"%15d                                            # Total wells\n",
-				Rxn_solution_map.size());
+				(int) Rxn_solution_map.size());
 
 		/* address */
 		fprintf(netpath_file,
